get from json function, tidying up
